chore: split classes into their own files
