+++ example/lib/widget_descriptor_tile.dart
@@ -1,0 +1,56 @@
+// Copyright 2017, Paul DeMarco.
+// All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:flutter/material.dart';
+import 'package:flutter_blue/flutter_blue.dart';
+
+class DescriptorTile extends StatelessWidget {
+  final BluetoothDescriptor descriptor;
+  final VoidCallback onReadPressed;
+  final VoidCallback onWritePressed;
+
+  const DescriptorTile(
+      {Key key, this.descriptor, this.onReadPressed, this.onWritePressed})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    var title = new Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        const Text('Descriptor'),
+        new Text(
+            '0x${descriptor.uuid.toString().toUpperCase().substring(4, 8)}',
+            style: Theme.of(context)
+                .textTheme
+                .body1
+                .copyWith(color: Theme.of(context).textTheme.caption.color))
+      ],
+    );
+    return new ListTile(
+      title: title,
+      subtitle: new Text(descriptor.value.toString()),
+      trailing: new Row(
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          new IconButton(
+            icon: new Icon(
+              Icons.file_download,
+              color: Theme.of(context).iconTheme.color.withOpacity(0.5),
+            ),
+            onPressed: onReadPressed,
+          ),
+          new IconButton(
+            icon: new Icon(
+              Icons.file_upload,
+              color: Theme.of(context).iconTheme.color.withOpacity(0.5),
+            ),
+            onPressed: onWritePressed,
+          )
+        ],
+      ),
+    );
+  }
+}
